feat: 사이드바 훅으로 분리
